Add ROI endpoint

+++ src/main/java/com/ibm/hybrid/cloud/sample/stocktrader/tradehistory/mongo/StockPurchase.java
@@ -50,7 +50,7 @@ public class StockPurchase implements Serializable{
         return when;
     }
 
-    public int getCommission() {
+    public double getCommission() {
         return commission;
     }
